requirements.txt
Django==5.0.6
djangorestframework==3.15.1
mysqlclient==2.2.4
Pillow==10.3.0
opencv-python==4.10.0.84
scikit-learn==1.5.0
django-cors-headers==4.4.0
google-api-python-client==2.137.0
djangorestframework-simplejwt==5.3.1
xhtml2pdf==0.2.16